requirements-dev.txt
autopep8==2.0.4
black==24.1.1
coverage==7.4.1
django-stubs==4.2.7
flake8==7.0.0
isort==5.13.2
mypy==1.7.0
pylint==3.0.3
pylint-django==2.5.5
pyyaml==6.0
typed-ast==1.5.5
types-Markdown==3.4.2.10
types-bleach==6.0.0.5
